Move Avatar Management to separate component. Add functionality for main user to change avatar.
